enable multiple load listeners for quicker load detection
Because the standard onStateChange observers are conservative in their decision as to whether a document is loaded, using other listeners, e.g. onReadyStateChange for a document, might speed up page display.
It might be possible to use the current onStateChange method and check the associated document with each fired event, to see if it's content has loaded (regardless of it's frames).