use weakref for all js objects
Link all objects to weak references _except a single pointer from JSClass to JSReference, and the pointers from outside the mozCom classes to the JSClass.
Example:
Clifox holds a hard reference to the current document.
That reference points to JSClass.
That JSClass has a pointer to it's backend .ref object.
That .ref object points back to the JSClass, as a weak reference.
Clifox gets rid of the current document.
The hard reference is deleted by Clifox deleting or change it to None.
The JSClass is deleted, calling it's own __del__ method in the process.
At this point, this class can notify the js side that it has been deleted, and can be removed from javascript as well.
This leaves the JSClass.ref hard reference dangling free, e.g. deleted, when the parent class is deleted.
Now, the JSReference is left allone, tethered to any other JSClass objects by weakrefs.
Weakrefs allow deletion when their are no hard references, so the JSRef class is deleted as well.