count refs for each object creation and deletion
We want to remove an object from the js map,
only when we no longer hold any references to it.
We can create multiple references pointing to the same object (think document.window, and window.self, etc).
The ID's of the objects will be the same, but the objects, in pythons view, will be different.
We make sure that we track the number of objects with the same id, and only remove objects whose count is 0.